+++ engine/src/graphics/renderer2d.hpp
@@ -15,9 +15,9 @@ class renderer2d : public frog::gx::renderer_base
 public:
     renderer2d() = default;
 
-    void viewport(int width, int height) override {}
+    void viewport(int /* width */, int /* height */) override {}
 
-    void clear_color(float r, float g, float b) override {}
+    void clear_color(float /* r */, float /* g */, float /* b */) override {}
 
     void clear() override {}
