+++ engine/src/lib2d/bits.hpp
@@ -1,0 +1,11 @@
+#pragma once
+
+#ifndef NOT_FROG_BUILD_2D
+
+#include "core.hpp"
+#include "structs.hpp"
+#include "events.hpp"
+#include "texture.hpp"
+#include "window.hpp"
+
+#endif
